infra/src/requirements.txt
aws-cdk.assets==1.42.0

aws-cdk.aws-cloudformation==1.42.0
aws-cdk.aws-codebuild==1.42.0
aws-cdk.aws-codecommit==1.42.0
aws-cdk.aws-codedeploy==1.42.0
aws-cdk.aws-codepipeline==1.42.0
aws-cdk.aws-codepipeline-actions==1.42.0
aws-cdk.aws_codestarconnections==1.42.0

aws-cdk.aws-iam==1.42.0
aws-cdk.aws-kms==1.42.0
aws-cdk.aws-lambda==1.42.0

aws-cdk.aws-s3==1.42.0
aws-cdk.aws-s3-assets==1.42.0
aws-cdk.aws-sns==1.42.0
aws-cdk.aws-sns-subscriptions==1.42.0

aws-cdk.aws-stepfunctions==1.42.0
aws-cdk.core==1.42.0
aws-cdk.aws-macie==1.42.0

